Refactor method_table to method_list (1d cell array)
Matlab tables are not a good fit after-all, they do not display nicely
on assignment: f = f.add_method(handle,types) does not print to screen
the actual content of method_table, you can see it nicely only if type
f.method_table afterwards. Also they are cumbersome to build.

So we move instead to a 1x2N cell, which displays nicely at assignment
and is quite simple to handle in the internals. There is a preliminary
attempt at a handle returning method but for now it can treat just one
input variable, so its useless. I have to figure how to get a function
handle that "dispatches" on varargin, the naive

handle = @(varargin)f(self,varargin)

does not work as expected.