Add failed test about dispatch on nargouts: we need to be more clear
Having this kind of powerful dispatch on nargouts would be so nice, but
I fear it could be impossible with Matlab, for how it treats calls with
omitted output arguments (just ignores them...). So if you want to stay
generic about how much stuff to return you should probably just conform
to the MathWorks way and write a unique specialized method for that, if
the case with some internal control on nargout. Since we support multi-
output returns that is indeed feasible and we are not limiting the use,
but still saying in the README that we can dispatch on nargout, without
further clarification can be quite misleading. We should probably write
extensively about this in the README.