First try at transforming interface from object class to function class
I actually don't like that much that the interface object cannot be called nicely and the handle is separated, it would make quite a bit for an awkward working API in a library that aims to ship generic functions. Maybe we should give up the upfront constructor and make it a proper ergonomic call to the dispatcher, so to ship only the interface object which could then be called or extended.

Something like

>> f = multimethod.interface;
>> f = f.add_method(fun1,...) ;
>> f = f.add_method(fun2,...);
>> f(varargin) % already dispatches

then the dispatch method would be private and for the end user f would just be some cool exotic magical thing.

Apparently this thing is called a functor, in C++ it is defined like this:

 // C++ program to demonstrate working of
 // functors.
 #include <bits/stdc++.h>
 using namespace std;

 // A Functor
 class increment
 {
 private:
     int num;
 public:
     increment(int n) : num(n) {  }

     // This operator overloading enables calling
     // operator function () on objects of increment
     int operator () (int arr_num) const {
         return num + arr_num;
     }
 };

// Driver code
int main()
{
    int arr[] = {1, 2, 3, 4, 5};
    int n = sizeof(arr)/sizeof(arr[0]);
    int to_add = 5;

    transform(arr, arr+n, arr, increment(to_add));

    for (int i=0; i<n; i++)
        cout << arr[i] << " ";
}

So what we need is actually a way to overload the '()' operator.

This thing looks promising: https://it.mathworks.com/help/matlab/ref/subsref.html

subsref is a special function name that allows overloading '()', '{}' and '.' operators, so I tried something like this:

    %% functor overload
    function varargout = subsref(obj,indexing)
    % >> varargout = multimethod_obj(varargin)
        if not(isequal(indexing.type,'()'))
            fprintf(2,"Error: braces and dots not supported for multimethod interfaces\n")
            [varargout{1:nargout}] = {};
            return
        end
        indexing.type
        indexing.subs{:}
        [varargout{:}] = obj.dispatch(indexing.subs{:});
    end

which DOES NOT work since I have no way for avoiding entering the method for dots, just to handle the error inside subsref.
> this breaks totally method calls and property accesses (the dot...).

We can then try a newer approach (an amend I guess) through: https://it.mathworks.com/help/matlab/ref/matlab.mixin.indexing.redefinesparen-class.html

But it will work only for ver >= R2021b (so let's think about it before merge?)